% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMP.R
\docType{data}
\name{BMP}
\alias{BMP}
\title{BMP image of phase map of cross section of modern oyster.}
\format{A BMP image imported into R as a large data array}
\source{
\url{https://doi.org/10.5194/gmd-2017-137-supplement}
}
\usage{
data(BMP)
}
\description{
An image of the microXRF phase map of the cross section used for this model
from the Crassostre gigas #1 oyster used as an example in de Winter (2017)
}
\keyword{datasets}
