% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_time.R
\docType{data}
\name{el_time}
\alias{el_time}
\title{Matrix listing the concentrations of elements in every subincrement}
\format{A data frame with 24 rows and 1291 variables:}
\source{
\url{https://doi.org/10.5194/gmd-2017-137-supplement}
}
\usage{
data(el_time)
}
\description{
A dataset containing the concentrations of every element measured in the
phase map of the XRF mapped surface of the Crassostrea gigas #1 oyster
used as an example in de Winter (2017) reconstructed in every subincrement.
}
\keyword{datasets}
