% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_length.R
\docType{data}
\name{image_length}
\alias{image_length}
\title{Length (in mm) of the digitized phase map image.}
\format{A single value:

\describe{
 Length in mm of digitized shell cross section
}}
\source{
\url{https://doi.org/10.5194/gmd-2017-137-supplement}
}
\usage{
data(image_length)
}
\description{
A single value of the image length in mm
}
\keyword{datasets}
