% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncG.R
\docType{data}
\name{IncG}
\alias{IncG}
\title{Interpolated subincrements in cross section of modern oyster rescaled to
the same X-axis.}
\format{A large data frame with 101 rows and 1291 variables:}
\source{
\url{https://doi.org/10.5194/gmd-2017-137-supplement}
}
\usage{
data(IncG)
}
\description{
A dataset containing X- and Y-coordinates of subincrements interpolated
between digitized growth increments from the Crassostrea gigas #1 oyster
used as an example in de Winter (2017) sampled to fit a common X-axis.
A Xstep of 0.1, a Tstep of 1 and a growth season of 250 days are used.
}
\keyword{datasets}
