% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incr_matrix1.R
\docType{data}
\name{incr_matrix1}
\alias{incr_matrix1}
\title{Matrix containing data calculated for each growth band.}
\format{A data frame with 8 rows and 5 variables:

\describe{
  \item{growth band}{name of the growth increment}
  \item{age (days)}{Age associated with the deposition of the growth increment}
  \item{age_cal (days)}{Age associated with the deposition of the growth increment, calibrated to the seasonal cycle}
  \item{incr_area}{area between subsequent increments}
  \item{incr_cumarea}{area between increment and the top of the shell}
}}
\source{
\url{https://doi.org/10.5194/gmd-2017-137-supplement}
}
\usage{
data(incr_matrix1)
}
\description{
A dataset containing specific parameters calculated for all digitized growth
increments from the Crassostre gigas #1 oyster used as an example in de Winter
(2017) sorted per increment.
}
\keyword{datasets}
