% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthfactor.R
\docType{data}
\name{lengthfactor}
\alias{lengthfactor}
\title{Multiplier used to convert shell cross section from pixels to mm}
\format{A single value:

\describe{
 Length in mm of one pixel in the digitized shell cross section
}}
\source{
\url{https://doi.org/10.5194/gmd-2017-137-supplement}
}
\usage{
data(lengthfactor)
}
\description{
A single value of the amount of mm contained in one pixel
}
\keyword{datasets}
