% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oyster_height.R
\docType{data}
\name{Oyster_height}
\alias{Oyster_height}
\title{Measured height of the shell}
\format{A single value:

\describe{
 Height in mm of the Crassostrea gigas #1 shell
}}
\source{
\url{https://doi.org/10.5194/gmd-2017-137-supplement}
}
\usage{
data(Oyster_height)
}
\description{
A single value containing the measured height in mm of the
Crassostrea gigas #1 oyster used as an example in de Winter
(2017) GMD.
}
\keyword{datasets}
