% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oyster_length.R
\docType{data}
\name{Oyster_length}
\alias{Oyster_length}
\title{Measured length of the shell}
\format{A single value:

\describe{
 Lenght in mm of the Crassostrea gigas #1 shell
}}
\source{
\url{https://doi.org/10.5194/gmd-2017-137-supplement}
}
\usage{
data(Oyster_length)
}
\description{
A single value containing the measured length (in 
anterio-posterior direction) in mm of the Crassostrea
gigas #1 oyster used as an example in de Winter (2017) GMD.
}
\keyword{datasets}
