% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_mat.R
\docType{data}
\name{phase_mat}
\alias{phase_mat}
\title{Matrix listing the amount of pixels of each phase in every subincrement}
\format{A data frame with 4 rows and 1291 variables:}
\source{
\url{https://doi.org/10.5194/gmd-2017-137-supplement}
}
\usage{
data(phase_mat)
}
\description{
A dataset containing the amounts of pixels of each phase in the
phase map of the XRF mapped surface of the Crassostrea gigas #1 oyster
used as an example in de Winter (2017) represented in every subincrement.
}
\keyword{datasets}
