% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_stat.R
\docType{data}
\name{phase_stat}
\alias{phase_stat}
\title{Statistics of elemental concentrations in XRF map of oyster}
\format{A data frame with 5 rows and 27 variables:

\describe{
  \item{Names}{Names of phases}
  \item{pixels}{Amount of pixels representing the phase}
  \item{fraction}{Fraction of map surface represented by phase}
  \item{C}{Concentration of C in phase}
  \item{O}{Concentration of O in phase}
  \item{Na}{Concentration of Na in phase}
  \item{Mg}{Concentration of Mg in phase}
  \item{Al}{Concentration of Al in phase}
  \item{Si}{Concentration of Si in phase}
  \item{P}{Concentration of P in phase}
  \item{S}{Concentration of S in phase}
  \item{Cl}{Concentration of Cl in phase}
  \item{K}{Concentration of K in phase}
  \item{Ca}{Concentration of Ca in phase}
  \item{Ti}{Concentration of Ti in phase}
  \item{Cr}{Concentration of Cr in phase}
  \item{Mn}{Concentration of Mn in phase}
  \item{Fe}{Concentration of Fe in phase}
  \item{Ni}{Concentration of Ni in phase}
  \item{Cu}{Concentration of Cu in phase}
  \item{Zn}{Concentration of Zn in phase}
  \item{Br}{Concentration of Br in phase}
  \item{Rb}{Concentration of Rb in phase}
  \item{Sr}{Concentration of Sr in phase}
  \item{Rh}{Concentration of Rh in phase}
  \item{Ba}{Concentration of Ba in phase}
  \item{Pb}{Concentration of Pb in phase}
}}
\source{
\url{https://doi.org/10.5194/gmd-2017-137-supplement}
}
\usage{
data(phase_stat)
}
\description{
A dataset containing trace element concentrations and fractions of a
phase map of the XRF mapped surface of the Crassostrea gigas #1 oyster
used as an example in de Winter (2017)
}
\keyword{datasets}
