% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasemat.R
\docType{data}
\name{phasemat}
\alias{phasemat}
\title{Matrix of phases names for each pixel in the phase map}
\format{A large data frame with 2258 rows and 2383 variables:}
\source{
\url{https://doi.org/10.5194/gmd-2017-137-supplement}
}
\usage{
data(phasemat)
}
\description{
A dataset containing names of phases for every pixel in the phase map
measured on a cross section through the Crassostrea gigas #1 oyster
used as an example in de Winter (2017)
}
\keyword{datasets}
