% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelsize.R
\docType{data}
\name{pixelsize}
\alias{pixelsize}
\title{Size of pixels in phase map in mm}
\format{A single value:

\describe{
 Length in mm of one pixel in the digitized shell cross section
 rounded up to the nearest micrometer
}}
\source{
\url{https://doi.org/10.5194/gmd-2017-137-supplement}
}
\usage{
data(pixelsize)
}
\description{
A single value of the amount of mm contained in one pixel
rounded up to the nearest micrometer.
}
\keyword{datasets}
