\name{ShellTrace}
\alias{ShellTrace}
\alias{ShellTrace}
\docType{package}
\title{ShellTrace: Growth and trace element uptake modelling in bivalve shells}
\description{This package contains formulae used to model the growth and
     development of bivalve shells based on digitized coordinated of shell
         increments in a longitudinal cross section thorugh the shell. The
             growth model is combined with XRF mapping results of the same
                 cross section and a seasonal growth rate model to model
                    trace element concentrations and uptake rates into the
                        bivalve shell.
                            de Winter, N. J. (2017) <doi:10.5194/gmd-2017-137>}
\details{Formulae in this package form the several steps of the model, and
     are not meant to be used individually. The order and application of
         these functions is outlined in the publication in Geoscientific
             Model Development that bears the name of the model (de Winter,
                 in review)}
\author{Niels J. de Winter

Maintainer: Niels J. de Winter
}
\references{de Winter, N. J.: ShellTrace v1.0 - A new approach for modelling
     growth and trace element uptake in marine bivalve shells: Model
         verification on pacific oyster shells (Crassostrea gigas), Geosci.
             Model Dev. Discuss., https://doi.org/10.5194/gmd-2017-137,
                 in review, 2017.}

\examples{print("de Winter, N. J.: ShellTrace v1.0 - A new approach for
     modelling growth and trace element uptake in marine bivalve shells:
         Model verification on pacific oyster shells (Crassostrea gigas),
             Geosci. Model Dev. Discuss., in review, 2017.")
}
