% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subincr_matrix2.R
\docType{data}
\name{subincr_matrix2}
\alias{subincr_matrix2}
\title{Matrix containing data calculated for each growth band.}
\format{A data frame with 1291 rows and 10 variables:

\describe{
  \item{Age}{age (in days) of the subincrement}
  \item{p1xs}{X-value of the first (leftmost) point in the subincrement}
  \item{p2xs}{X-value of the last (rightmost) point in the subincrement}
  \item{areaY}{Area between subsequent subincrements}
  \item{areaC}{Area between subincrement and top of shell}
  \item{p1y}{Y-value of the first (leftmost) point in the subincrement}
  \item{p2y}{Y-value of the last (rightmost) point in the subincrement}
  \item{shell_height}{Height of shell during deposition of the subincrement}
  \item{firstl}{Row number in IncG of first (leftmost) data point belonging to the subincrement}
  \item{lastl}{Row number in IncG of last (rightmost) data point belonging to the subincrement}
}}
\source{
\url{https://doi.org/10.5194/gmd-2017-137-supplement}
}
\usage{
data(subincr_matrix2)
}
\description{
A dataset containing specific parameters calculated for all interpolated
subincrements from the Crassostrea gigas #1 oyster used as an example in de Winter
(2017) sorted per increment.
}
\keyword{datasets}
