% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z_mat.R
\docType{data}
\name{Z_mat}
\alias{Z_mat}
\title{Z-values describing the base ellipse of the oyster}
\format{A large data frame with 101 rows and 1291 variables:}
\source{
\url{https://doi.org/10.5194/gmd-2017-137-supplement}
}
\usage{
data(Z_mat)
}
\description{
A dataset containing Z-coordinates of the base ellipse calculated
for all subincrements in the Crassostrea gigas #1 oyster used as
an example in de Winter (2017).
A Xstep of 0.1, a Tstep of 1 and a growth season of 250 days are used.
}
\keyword{datasets}
